+++ docs/screenshots/command_line_top_screenshot.py
@@ -1,0 +1,10 @@
+"""Screenshot to show the command line on top."""
+
+from support.maker import make_app
+
+app = make_app(command_line_on_top=True)
+
+if __name__ == "__main__":
+    app.run()
+
+### command_line_top_screenshot.py ends here